+++ klippy/chelper/kin_delta.c
@@ -10,6 +10,7 @@
 #include <string.h> // memset
 #include "compiler.h" // __visible
 #include "itersolve.h" // struct stepper_kinematics
+#include "trapq.h" // move_get_coord
 
 struct delta_stepper {
     struct stepper_kinematics sk;
